# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PMCI API
  version: "2026-03-02"
  description: |
    Prediction Market Canonical Intelligence API.

    Observe cross-provider market prices, divergence signals, coverage gaps, and manage
    market-link review proposals. Observation-only — no trading.

    **Auth tiers:**
    - **Public** — `/v1/health/*` routes, no key required.
    - **Standard** — All other routes. Send `x-pmci-api-key` when `PMCI_API_KEY` is configured on the server.
    - **Admin** — `POST /v1/resolve/link` only. Send `x-pmci-admin-key` when `PMCI_ADMIN_KEY` is configured.

    When neither env var is set the API is fully open — useful for local development.
servers:
  - url: http://localhost:8787
    description: Local development (no TLS)
  - url: https://{host}
    description: Production (Caddy TLS termination)
    variables:
      host:
        default: api.yourdomain.com
        description: Value of PMCI_HOST env var
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: x-pmci-api-key
      description: |
        Required on all non-health routes when `PMCI_API_KEY` env var is set on the server.
        Returns 401 if the key is wrong or missing.
    AdminKeyHeader:
      type: apiKey
      in: header
      name: x-pmci-admin-key
      description: |
        Required only for `POST /v1/resolve/link` when `PMCI_ADMIN_KEY` env var is set on the server.
  headers:
    XPmciVersion:
      description: API version string. Present on every response.
      schema:
        type: string
        example: "2026-03-02"
  schemas:
...
    ErrorUnauthorized:
      type: object
      required: [error]
      properties:
        error:
          type: string
          enum: [unauthorized]
    ErrorRateLimited:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          enum: [rate_limited]
        message:
          type: string
          example: "Too many requests. See max and timeWindow in response headers."
    ErrorStale:
      type: object
      required: [error]
      properties:
        error:
          type: string
          enum: [stale_data]
        lag_seconds:
          type: integer
          nullable: true
        max_lag_seconds:
          type: integer
        message:
          type: string
...
    DivergenceSignal:
      type: object
      properties:
        family_id:
          type: integer
        provider:
          type: string
        provider_market_id:
          type: integer
        relationship_type:
          type: string
        price:
          type: number
          nullable: true
        consensus_price:
          type: number
          nullable: true
        divergence:
          type: number
          nullable: true
    TopDivergenceResult:
      type: object
      properties:
        family_id:
          type: integer
        label:
          type: string
        consensus_price:
          type: number
          nullable: true
        max_divergence:
          type: number
          nullable: true
        last_observed_at:
          type: string
          format: date-time
          nullable: true
        legs:
          type: array
          items:
            type: object
            properties:
              provider:
                type: string
              provider_market_id:
                type: integer
              provider_market_ref:
                type: string
                nullable: true
              price_yes:
                type: number
                nullable: true
              divergence:
                type: number
                nullable: true
              relationship_type:
                type: string
              confidence:
                type: number
...
  responses:
    Unauthorized:
      description: API key missing or incorrect.
      headers:
        X-PMCI-Version:
          $ref: "#/components/headers/XPmciVersion"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorUnauthorized"
    RateLimited:
      description: Rate limit exceeded. Default is 60 requests per 60 seconds per key.
      headers:
        X-PMCI-Version:
          $ref: "#/components/headers/XPmciVersion"
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorRateLimited"
    Stale:
      description: |
        Data is stale — observer lag exceeds `PMCI_MAX_LAG_SECONDS` (default 120s).
        Only returned from `/v1/signals/*` routes. Restart the observer with `npm run start`.
      headers:
        X-PMCI-Version:
          $ref: "#/components/headers/XPmciVersion"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorStale"
paths:
...
  /v1/signals/divergence:
    get:
      summary: Per-link divergence for a family
      description: |
        Returns the current price divergence for each link in a family. Requires fresh data —
        returns `503` if the observer lag exceeds `PMCI_MAX_LAG_SECONDS`.
      operationId: getDivergence
      tags: [Signals]
      security:
        - ApiKeyHeader: []
        - {}
      parameters:
        - name: family_id
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Divergence signals
          headers:
            X-PMCI-Version:
              $ref: "#/components/headers/XPmciVersion"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DivergenceSignal"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/Stale"
  /v1/signals/top-divergences:
    get:
      summary: Top divergences across an event
      description: |
        Returns the families with the highest cross-provider price divergence under a canonical event,
        sorted by `max_divergence` descending. Requires fresh data — returns `503` if stale.
      operationId: getTopDivergences
      tags: [Signals]
      security:
        - ApiKeyHeader: []
        - {}
      parameters:
        - name: event_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Top divergence results
          headers:
            X-PMCI-Version:
              $ref: "#/components/headers/XPmciVersion"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopDivergenceResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/Stale"
  /v1/snapshots:
    get:
      summary: Price history for a market family
      description: |
        Returns bucketed average `price_yes` per provider for all active links in a family.
        Use to draw sparklines on the candidate detail view. No freshness gate — returns
        historical data even if the observer is temporarily stale.
      operationId: getSnapshots
      tags: [Signals]
      security:
        - ApiKeyHeader: []
        - {}
      parameters:
        - name: family_id
          in: query
          required: true
          schema:
            type: integer
        - name: since
          in: query
          required: false
          description: How far back to look. Format `{N}h` or `{N}d` (e.g. `7d`, `24h`).
          schema:
            type: string
            default: 7d
            example: 7d
        - name: interval
          in: query
          required: false
          description: Bucket size for time-series aggregation.
          schema:
            type: string
            enum: [1h, 1d]
            default: 1h
      responses:
        "200":
          description: Bucketed price history grouped by provider
          headers:
            X-PMCI-Version:
              $ref: "#/components/headers/XPmciVersion"
          content:
            application/json:
              schema:
                type: object
                properties:
                  family_id:
                    type: integer
                  interval:
                    type: string
                    enum: [1h, 1d]
                  since:
                    type: string
                  providers:
                    type: array
                    items:
                      type: object
                      properties:
                        provider:
                          type: string
                          example: kalshi
                        series:
                          type: array
                          items:
                            type: object
                            properties:
                              bucket:
                                type: string
                                format: date-time
                              price_yes:
                                type: number
                                format: float
              example:
                family_id: 42
                interval: 1h
                since: 7d
                providers:
                  - provider: kalshi
                    series:
                      - bucket: "2026-03-10T00:00:00Z"
                        price_yes: 0.54
                      - bucket: "2026-03-10T01:00:00Z"
                        price_yes: 0.55
                  - provider: polymarket
                    series:
                      - bucket: "2026-03-10T00:00:00Z"
                        price_yes: 0.46
                      - bucket: "2026-03-10T01:00:00Z"
                        price_yes: 0.47
        "400":
          description: Invalid parameters
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"

  # ── Review ─────────────────────────────────────────────────────────────────
...